api/requirements.txt


fastapi == 0.104.1
uvicorn == 0.24.0.post1 

python-jose[Cyptograph] == 3.3.0
python-multipart == 0.0.6
motor == 3.3.2
passlib[bcrypt] == 1.7.4